korekto por 2043f6d (403 erarpaĝo kun kromaj detaloj)
